[Update Lib] - Handle confirm password and Clean code.
